FACES-1566 Add ResourceHandlerBase and refactor ResourceHandlers implementing handleResourceRequest method
